+++ Sources/Utilities/KMCellConfiguration.swift
@@ -1,0 +1,14 @@
+//
+//  KMCellConfiguration.swift
+//  KommunicateChatUI-iOS-SDK
+//
+//  Created by sathyan elangovan on 22/04/22.
+//
+
+import Foundation
+/// - NOTE: Customization for message cells
+public enum KMCellConfiguration {
+    
+    /// if true then incoming messsage's sender name will be hidden on conversation
+    public static var hideSenderName = false
+}