+++ Sources/Views/ALKFriendLocationCell.swift
@@ -11,7 +11,6 @@ import UIKit
 
 final class ALKFriendLocationCell: ALKLocationCell {
     // MARK: - Declare Variables or Types
-
     // MARK: Environment in chat
 
     private var avatarImageView: UIImageView = {
@@ -58,6 +57,7 @@ final class ALKFriendLocationCell: ALKLocationCell {
 
         timeLabel.leadingAnchor.constraint(equalTo: bubbleView.trailingAnchor, constant: 2.0).isActive = true
         timeLabel.bottomAnchor.constraint(equalTo: bubbleView.bottomAnchor, constant: 2.0).isActive = true
+        nameLabel.isHidden = KMCellConfiguration.hideSenderName
     }
 
     override func setupStyle() {