+++ Sources/Views/ALKMessageButtonCell.swift
@@ -356,6 +356,7 @@ class ALKFriendMessageButtonCell: ALKChatBaseCell<ALKMessageViewModel> {
             timeLabelHeight,
             timeLabel.trailingAnchor.constraint(lessThanOrEqualTo: trailingAnchor),
         ])
+        nameLabel.isHidden = KMCellConfiguration.hideSenderName
     }
 }